# openapi/openapi.yaml
...
openapi: "3.0.0"
info:
  version: 0.1.0
  title: Galaxy API
  license:
    name: Apache-2.0
  description: |
    # Introduction

    This is the <a href="https://www.openapis.org">OpenAPI</a> specification
    for the **Red Hat Ansible Automation Hub** API.

    ## OpenAPI Information
    <a target="_top" href="https://swagger.io/docs/specification/about/">Swagger.io OpenAPI documentation</a>

    <a target="_top" href="https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md">OpenAPI Specification</a>

    <a target="_top" href="https://github.com/ansible/galaxy-api/blob/master/openapi/openapi.yaml">Download the Automation Hub OpenAPI Specification</a>.


    # API Conventions
    ## Pagination

    Resource lists are always paginated.

    This pagination format is based on the
    <a href="https://jsonapi.org/format/#document-structure">JSON API document
    structure.</a>

     The response will look similar to this:

    ```json

    {
      "meta": {
        "count": 5
      },
      "links": {
        "first": "/namespaces/?offset=0&limit=1",
        "last": "/namespaces/?offset=7&limit=1",
        "next": "/namespaces/?offset=5&limit=1",
        "previous": "/namespaces/?offset=3&limit=1"
      },
      "data": [
        {
          "name": "my_namespace"
        }
      ]
    }
    ```

    ## Errors

    Error responses always have a top level object with an 'errors' field.

    The value of the 'errors' field is a list of Error objects.

...

...
      {
...
servers:
...
paths:
...
  '/v3/collections/':
    get:
      summary: List Collections
      operationId: listCollections
      parameters:
        - $ref: '#/components/parameters/PageOffset'
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/SearchKeyword'
        - $ref: '#/components/parameters/SearchName'
        - $ref: '#/components/parameters/SearchNamespace'
        - $ref: '#/components/parameters/SearchTag'
      tags:
        - Collections
      responses:
        '200':
          $ref: '#/components/responses/CollectionList'
        'default':
          $ref: '#/components/responses/Errors'
  '/v3/collections/{namespace}/{name}/':
    parameters:
      - $ref: '#/components/parameters/CollectionNamespaceName'
      - $ref: '#/components/parameters/CollectionName'
    get:
      summary: Get Collection
      operationId: getCollection
      tags:
        - Collections
      responses:
        '200':
          $ref: '#/components/responses/Collection'
        'default':
          $ref: '#/components/responses/Errors'
...
  '/v3/artifacts/collections/':
    post:
      summary: Upload Collection Artifact
      operationId: uploadCollectionArtifact
      tags:
        - Artifacts
      requestBody:
        $ref: '#/components/requestBodies/ArtifactUpload'
      responses:
        '202':
          $ref: '#/components/responses/CollectionImportAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Errors'
  '/v3/artifacts/collections/{filename}':
    get:
      summary: Download Collection Artifact
      operationId: downloadCollectionArtifact
      tags:
        - Artifacts

      parameters:
        - $ref: '#/components/parameters/CollectionVersionArtifactFilename'
      responses:
        '200':
          description: 'A requested artifact file.'
          content:
            application/octet-stream:
              schema:
                description: 'The collection artifact file binary contents.'
                type: string
                format: binary
        'default':
          $ref: '#/components/responses/Errors'
  '/content/{distribution_base_path}/v3/artifacts/collections/':
    parameters:
      - $ref: '#/components/parameters/DistributionBasePath'
    post:
      summary: Upload Collection Artifact To A Repository and Distribution
      operationId: uploadCollectionArtifactToRepository
      tags:
        - Artifacts
        - PerDistribution
      requestBody:
        $ref: '#/components/requestBodies/ArtifactUpload'
      responses:
        '202':
          $ref: '#/components/responses/CollectionImportAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Errors'
  '/content/{distribution_base_path}/v3/collections/':
    parameters:
      - $ref: '#/components/parameters/DistributionBasePath'
    get:
      summary: List Collections In a Repository and Distribution
      operationId: listCollectionsInARepository
      parameters:
        - $ref: '#/components/parameters/PageOffset'
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/SearchKeyword'
        - $ref: '#/components/parameters/SearchName'
        - $ref: '#/components/parameters/SearchNamespace'
        - $ref: '#/components/parameters/SearchTag'
      tags:
        - Collections
        - PerDistribution
      responses:
        '200':
          $ref: '#/components/responses/CollectionList'
        'default':
          $ref: '#/components/responses/Errors'
  '/content/{distribution_base_path}/v3/collections/{namespace}/{name}/':
    parameters:
      - $ref: '#/components/parameters/DistributionBasePath'
      - $ref: '#/components/parameters/CollectionNamespaceName'
      - $ref: '#/components/parameters/CollectionName'
    get:
      summary: Get Collection In a Repository and Distribution
      operationId: getCollectionInARepository
      tags:
        - Collections
        - PerDistribution
      responses:
        '200':
          $ref: '#/components/responses/Collection'
        'default':
          $ref: '#/components/responses/Errors'
...
  '/v3/imports/collections/{import_id}/':
    get:
      summary: Get Collection Import
      operationId: getCollectionImport
      parameters:
        - $ref: '#/components/parameters/CollectionImportId'
      tags:
        - Imports
      responses:
        '200':
          $ref: '#/components/responses/CollectionImport'
        'default':
          $ref: '#/components/responses/Errors'
...
  '/content/{distribution_base_path}/v3/imports/collections/{import_id}/':
    parameters:
      - $ref: '#/components/parameters/DistributionBasePath'
    get:
      summary: Get Collection Import In A Repository And Distribution
      operationId: getCollectionImportInARepository
      parameters:
        - $ref: '#/components/parameters/CollectionImportId'
      tags:
        - Imports
        - PerDistribution
      responses:
        '200':
          $ref: '#/components/responses/CollectionImport'
        'default':
          $ref: '#/components/responses/Errors'
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic

    TokenAuth:
      type: apiKey
      name: Authorization
      in: header
      description: |
        Django Rest Framework (DRF) token authentication.
        Passed in HTTP header as:
          Authorization: Token <token>
  schemas:
...
    Collection:
      title: 'Collection'
      description: 'Ansible content collection'
      type: object
      properties:
        name:
          description: 'Collection name. Must be lower case containing
                        only alphanumeric characters and underscores.'
          type: string
          maxLength: 64
          readOnly: true
        namespace:
          $ref: '#/components/schemas/NamespaceSummary'
        latest_version:
          $ref: '#/components/schemas/CollectionVersionLink'
        created:
          type: string
          readOnly: true
        deprecated:
          type: boolean
        modified:
          type: string
          readOnly: true
        id:
          type: integer
          readOnly: true
        href:
          type: string
          format: uri
        versions_url:
          type: string
      required:
        - name
        - remote_id
        - namespace
    CollectionCreationResult:
      title: 'CollectionCreationResult'
      description: 'A map of collection import task info, including its url'
      type: object
      properties:
        task:
          description: >
            The url for the queued collection import task.
            Check it for progress and status.
          type: string
          format: uri
    CollectionImport:
      description: 'Collection import'
      title: 'Collection Import'
      type: object
      properties:
        error:
          $ref: '#/components/schemas/CollectionImportError'
        finished_at:
          type: string
          format: date-time
        id:
          type: integer
        imported_version:
          type: string
          readOnly: true
        job_id:
          type: string
          format: uuid
        messages:
          type: array
          items:
            $ref: '#/components/schemas/CollectionImportMessage'
        name:
          maxLength: 64
          type: string
        namespace:
          type: string
        started_at:
          type: string
          format: date-time
        state:
          type: string
        version:
          maxLength: 64
          type: string
      required:
        - error
        - finished_at
        - id
        - job_id
        - messages
        - name
        - namespace
        - started_at
        - state
        - version
    CollectionImportError:
      description: 'Error reported from collection importer'
      title: Collection Import Error
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        traceback:
          type: string
    CollectionImportMessage:
      description: 'Message from collection importer'
      title: Collection Import Message
      type: object
      properties:
        level:
          type: string
        message:
          type: string
        time:
          type: string
          format: date-time
      required:
        - level
        - message
        - time
...
    CollectionsPage:
      description: 'A page of a list of Collections'
      allOf:
        - $ref: '#/components/schemas/PageInfo'
        - type: object
          properties:
            data:
              description: 'List of Collections for this Page'
              title: 'Collections'
              type: array
              items:
                $ref: '#/components/schemas/Collection'
          required:
            - data

    # -------------------------------------
    # Schemas: Collection
    # -------------------------------------
...
    CollectionVersionLink:
      title: 'Collection Version Link'
      type: object
      properties:
        href:
          description: 'The URL to the CollectionVersion'
          type: string
        version:
          $ref: '#/components/schemas/SemanticVersion'
...
    Errors:
      title: 'Errors'
      description: "A list of JSON API Error objects"
      type: object
      externalDocs:
        description: 'JSON API Error Specification'
        url: 'https://jsonapi.org/format/#errors'
      properties:
        errors:
          type: array
          title: 'Errors'
          items:
            $ref: '#/components/schemas/Error'
          minItems: 1
      required:
        - errors
    Error:
      title: 'Error'
      description: "A JSON API Error object"
      externalDocs:
        description: 'JSON API Error Specification'
        url: 'https://jsonapi.org/format/#error-objects'
      type: object
      properties:
        code:
          description: 'Unique identifier for the error'
          type: string
          example: 'not_found'
        detail:
          type: string
          description: >
            A human-readable explanation specific to this
            occurrence of the problem
          example: 'Record /unicorn/11 was not found'
        title:
          type: string
          description: 'A short, human-readable summary of the problem'
          example: 'Record not found.'
        status:
          type: string
          description: 'String representation of HTTP status code'
          example: '404'
        source:
          type: object
          properties:
            parameter:
              description: >
                A string indicating which URI query parameter
                caused the error.
              type: string
            pointer:
              description: >
                A JSON Pointer [RFC6901] to the associated entity
                in the request document
              type: string
      required:
        - detail
        - status
...
    NamespaceName:
      title: 'NamespaceName'
      description: >
          Namespace name. Must be lower case containing
          only alphanumeric characters and underscores.
      type: string
      maxLength: 64
      example: 'my_namespace'
      pattern: '^(?!.*__)[a-z]+[0-9a-z_]*$'
...
    NamespaceSummary:
      type: object
      title: 'Namespace Summary'
      properties:
        name:
          $ref: '#/components/schemas/NamespaceName'
        id:
          type: integer
        href:
          description: 'Link to the Namespace'
          type: string
          format: uri

    # -------------------------------------
    # Schemas: Pagination
    # -------------------------------------
    PageInfo:
      description: 'Pagination info'
      title: 'Page Info'
      properties:
        meta:
          $ref: '#/components/schemas/PageMeta'
        links:
          $ref: '#/components/schemas/PageLinks'
      required:
        - meta
        - links
...
    PageMeta:
      title: 'PageMeta'
      type: object
      properties:
        count:
          description: 'The number of items in this Page'
          type: integer
          example: 5
      required:
        - count
...
  parameters:
    CollectionNamespaceName:
      description: 'The collection namespace name'
      in: path
      name: namespace
      required: true
      schema:
        $ref: '#/components/schemas/NamespaceName'
    CollectionName:
      description: 'The collection name'
      in: path
      name: name
      required: true
      schema:
        type: string
        pattern: ^(?!.*__)[a-z]+[0-9a-z_]*$'
    CollectionVersionArtifactFilename:
      description: 'CollectionVersion artifact filename'
      in: path
      name: filename
      required: true
      schema:
        type: string
      example: 'testing.nginx-1.2.3.tar.gz'
    CollectionImportId:
      description: 'A unique integer value identifying a collection import.'
      in: path
      name: import_id
      required: true
      schema:
        type: string
...
    DistributionBasePath:
      description: 'The base_path of the pulp Distribution'
      in: path
      name: distribution_base_path
      required: true
      schema:
        type: string
      example: 'automation-hub'
    PageLimit:
      description: 'Number of results to return per page.'
      in: query
      name: limit
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
    PageOffset:
      description: 'Page offset number within the paginated result set'
      in: query
      name: offset
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
...
    SearchKeyword:
      description: 'Search for a keyword'
      in: query
      name: keywords
      required: false
      schema:
        type: string
    SearchName:
      description: 'Search for a name'
      in: query
      name: names
      required: false
      schema:
        type: string
    SearchNamespace:
      description: 'Search for a namespace'
      in: query
      name: namespaces
      required: false
      schema:
        type: string
    SearchTag:
      description: 'Search for a tag'
      in: query
      name: tags
      required: false
      schema:
        type: string
...
  requestBodies:
    ArtifactUpload:
      content:
        multipart/form-data:
          schema:
            description: >
              A multipart/form encoded payload including the binary
              collection artifact file contents and it's sha256 checksum.
            type: object
            properties:
              sha256:
                description: 'The sha256 digest of the collection artifact file'
                type: string
              file:
                description: 'The binary contents of a collection artifact'
                type: string
                format: binary
            required:
              - file
...
  responses:
    Collection:
      description: 'Response containing a Collection'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Collection'
    CollectionList:
      description: 'Response containing a page of Collections'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CollectionsPage'
...
    CollectionImportAccepted:
      description: >
        Result of an accepted collection import.
        Includes the url of the import task
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CollectionCreationResult'
...
    Conflict:
      description: 'Conflict Error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
    CollectionImport:
      description: The requested Collection Import
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CollectionImport'
...
    Errors:
      description: 'Errors'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
...
    Unauthorized:
      description: 'Unauthorized (401)'
      headers:
        WWW-Authenticate:
          schema:
            type: string
            enum:
              - Basic
              - Token
              - Bearer
          description: >
            Authentication Types:
             * `Basic` - HTTP Basic Auth
             * `Token` - Django style auth token, used with Galaxy API v2.
             * `Bearer` - Bearer token, used with Galaxy API v3
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
...